add and del user from table users
