+++ Server/inc/Chat.h
@@ -37,11 +37,13 @@ int *client_anch, *client_ash;
 ///////
 char *mx_autentification(int sock);
 void *connection_handler(void *new_sock);
+
 ////
 char *mx_strnew(const int size);
 int mx_strlen(const char *s);
 char *mx_strcat(char *restrict s1, const char *restrict s2);
 char *mx_strjoin(char const *s1, char const *s2);
+char* int_to_str(int num);
 
 ////Database
 extern sqlite3* db;
@@ -52,6 +54,7 @@ void db_exec(char* statement, sqlite3* db);
 void db_add_user_to_online(char *login, int socket, sqlite3* db);
 void db_del_user_from_online(char *login, sqlite3* db);
 int db_get_online_user_socket(char *login, sqlite3* db);
+void db_add_user(char *login, char *password, char *name);
+void db_del_user(char *login);
 
-char* int_to_str(int num);
 #endif
